requirements.txt
# Core dependencies
telethon>=1.32.0
google-generativeai>=0.3.0
requests>=2.31.0

# Optional: For better development experience
python-dotenv>=1.0.0

# Note: Some packages are built-in with Python and don't need installation:
# - asyncio (Python 3.7+)
# - sqlite3 (built-in)
# - pathlib (Python 3.4+)
# - dataclasses (Python 3.7+)
# - typing (Python 3.5+)
# - json (built-in)
# - logging (built-in)
# - datetime (built-in)

# System package needed (not installable via pip):
 python3-tk  